Refactor postorders to its own util (#61)
